Added write handler functionality, tuned the trap event detection, changed state display to use LEDs

+++ src/peripheral/LIS2DH12.h
@@ -186,10 +186,10 @@ extern LIS2DH12_Ret LIS2DH12_getALLmG(int32_t* const accX, int32_t* const accY, 
 
 extern void LIS2DH12_clearInterrupts();
 
-extern void LIS2DH12_setInterruptThreshold(uint8_t threshold);
+extern void LIS2DH12_setInterruptThreshold(uint16_t threshold);
 extern void LIS2DH12_setInterruptDuration(uint8_t duration);
 extern void LIS2DH12_setInterruptHandler(gpio_event_handler_t handler);
-extern void LIS2DH12_initThresholdInterrupt(uint8_t threshold,
+extern void LIS2DH12_initThresholdInterrupt(uint16_t threshold,
                                              uint8_t duration,
                                              LIS2DH12_InterruptThresholdMask intThreshMask,
                                              bool latchEnabled,
